Moved leaderboard to it's own package
