+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/CompetitionType.java
@@ -2,6 +2,10 @@ package com.oheers.fish.competition;
 
 public enum CompetitionType {
 
-    LARGEST_FISH, SPECIFIC_FISH, MOST_FISH, SPECIFIC_RARITY, LARGEST_TOTAL,
+    LARGEST_FISH,
+    SPECIFIC_FISH,
+    MOST_FISH,
+    SPECIFIC_RARITY,
+    LARGEST_TOTAL,
     RANDOM
 }
