+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/leaderboard/Leaderboard.java
@@ -1,43 +1,13 @@
-package com.oheers.fish.competition;
+package com.oheers.fish.competition.leaderboard;
 
+import com.oheers.fish.competition.CompetitionEntry;
+import com.oheers.fish.competition.CompetitionType;
 import com.oheers.fish.fishing.items.Fish;
 
 import java.util.Iterator;
 import java.util.Set;
 import java.util.TreeSet;
 import java.util.UUID;
-
-interface LeaderboardHandler {
-
-    Set<CompetitionEntry> getEntries();
-
-    void addEntry(UUID player, Fish fish);
-
-    void addEntry(CompetitionEntry entry);
-
-    void clear();
-
-    boolean contains(CompetitionEntry entry);
-
-    CompetitionEntry getEntry(UUID player);
-
-    Iterator<CompetitionEntry> getIterator();
-
-    int getSize();
-
-    boolean hasEntry(UUID player);
-
-    void removeEntry(CompetitionEntry entry) throws Exception;
-
-    CompetitionEntry getTopEntry();
-
-    UUID getPlayer(int i);
-
-    float getPlaceValue(int i);
-
-    Fish getPlaceFish(int i);
-
-}
 
 public class Leaderboard implements LeaderboardHandler {
