+++ even-more-fish-plugin/src/main/java/com/oheers/fish/database/DatabaseV3.java
@@ -4,7 +4,7 @@ import com.devskiller.friendly_id.FriendlyId;
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.competition.Competition;
 import com.oheers.fish.competition.CompetitionEntry;
-import com.oheers.fish.competition.Leaderboard;
+import com.oheers.fish.competition.leaderboard.Leaderboard;
 import com.oheers.fish.database.connection.ConnectionFactory;
 import com.oheers.fish.database.connection.MySqlConnectionFactory;
 import com.oheers.fish.database.connection.SqliteConnectionFactory;
